Conduct sparse mediation with elastic net (Old version)

Fit a mediation model via penalized maximum likelihood and structural equation model. The regularization path is computed for the lasso or elasticnet penalty at a grid of values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.

sparse.mediation.old(X, M, Y, tol = 10^(-10), max.iter = 100,
  lambda = log(1 + (1:50)/125), glmnet.penalty.factor = c(0, rep(1, 2 *
  V)), alpha = 1, tau = 1, verbose = FALSE)

Arguments

X
One-dimensional predictor

M
Multivariate mediator

Y
Outcome

tol
(default -10^(-10)) convergence criterion

max.iter
(default=100) maximum iteration

lambda
(default=log(1+(1:50)/125)) tuning parameter for L1 penalization

glmnet.penalty.factor
(default=c(0,rep(1,2*V))) give different weight of penalization for the 2V mediation paths.

alpha
(defult=1) tuning parameter for L2 penalization

tau
(default=1) tuning parameter for L1 penality weighting for paths a and b.

verbose
(default=FALSE) print progress

Value

c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths

Details

Multiple Mediaton Model: (1) M = Xa + e1 (2) Y = Xc' + Mb + e2 And in the optimization, we do not regularize c', due to the assumption of partial mediation.

References

TBA

Examples

N=100
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X %*% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M %*% b + rnorm(N)
sparse.mediation.old(X,M,Y,tol=10^(-10),
max.iter=100,lambda = log(1+(1:25)/50))

#> $c
#>  [1] 0.04022092 0.03600612 0.03433900 0.03243881 0.03085479 0.02963070
#>  [7] 0.02789994 0.02636909 0.02587669 0.02538428 0.02548996 0.02523398
#> [13] 0.02536482 0.02540086 0.02580411 0.02588929 0.02629656 0.02613293
#> [19] 0.02631453 0.02659668 0.02653930 0.02680362 0.02706866 0.02706577
#> [25] 0.02707630
#> 
#> $hatb
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,] -0.0008486792  0.000000000  0.000000000  0.000000000  0.000000000
#>  [2,] -0.2276407745 -0.178476122 -0.157280695 -0.138957760 -0.117265113
#>  [3,]  0.1242605695  0.101562788  0.080487407  0.055109273  0.029575482
#>  [4,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#>  [5,] -0.1708180479 -0.133037747 -0.115795287 -0.092549551 -0.065780271
#>  [6,] -0.0077368105  0.000000000  0.000000000  0.000000000  0.000000000
#>  [7,]  0.0743262602  0.050466575  0.027506278  0.014773473  0.000000000
#>  [8,] -0.0198411043 -0.008123633  0.000000000  0.000000000  0.000000000
#>  [9,] -0.0240009731 -0.036665997 -0.042265782 -0.035963528 -0.027850172
#> [10,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [11,]  0.0602117803  0.018528816  0.000000000  0.000000000  0.000000000
#> [12,] -0.2038213463 -0.151566324 -0.111915835 -0.077935894 -0.043435124
#> [13,]  0.0816750324  0.045378831  0.008601991  0.000000000  0.000000000
#> [14,]  5.1221367831  5.110657168  5.094139305  5.094910094  5.101041029
#> [15,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [16,] -0.0543430855 -0.006652661  0.000000000  0.000000000  0.000000000
#> [17,] -0.1412639379 -0.136303557 -0.128775910 -0.126761022 -0.120530938
#> [18,]  0.0819311317  0.007174270  0.000000000  0.000000000  0.000000000
#> [19,] -0.1676223477 -0.141432376 -0.130046202 -0.119070504 -0.105366700
#> [20,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [21,] -0.0205008301 -0.004695741 -0.001265497  0.000000000  0.000000000
#> [22,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [23,] -0.1506853704 -0.116434195 -0.096272028 -0.086018142 -0.079039868
#> [24,] -0.1096924147 -0.072887317 -0.048180022 -0.027534952 -0.008736537
#> [25,]  0.0181860359  0.006547161  0.004256775  0.000000000  0.000000000
#> [26,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [27,]  0.0702872349  0.042495104  0.019146740  0.001465286  0.000000000
#> [28,]  4.7811948292  4.823968958  4.842266440  4.854190025  4.864892548
#> [29,] -0.0673889073 -0.049639646 -0.025881641 -0.006914217  0.000000000
#> [30,] -0.0480967404 -0.029900507 -0.013954231  0.000000000  0.000000000
#> [31,]  0.0882210991  0.034063512  0.005072810  0.000000000  0.000000000
#> [32,]  0.1441720954  0.107659248  0.071461852  0.051234435  0.033859088
#> [33,]  0.0477368546  0.044058907  0.031202505  0.012520573  0.000000000
#> [34,] -0.0258833420  0.000000000  0.000000000  0.000000000  0.000000000
#> [35,]  0.1886374245  0.151999773  0.120738250  0.101166827  0.084773588
#> [36,] -0.0888157367 -0.036681879 -0.007368479  0.000000000  0.000000000
#> [37,]  0.0476724361  0.018407922  0.003399059  0.000000000  0.000000000
#> [38,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [39,]  4.6937331487  4.719880896  4.740183045  4.754512414  4.759395648
#> [40,]  0.1165837318  0.076609729  0.046729921  0.022859227  0.003967662
#> [41,] -0.2023502469 -0.158519700 -0.132725334 -0.100692592 -0.070585637
#> [42,] -0.1003168426 -0.066758372 -0.057102582 -0.045565709 -0.026519852
#> [43,] -0.1047137059 -0.064889373 -0.056161757 -0.044333816 -0.028280397
#> [44,] -0.0706958716 -0.026401012  0.000000000  0.000000000  0.000000000
#> [45,]  0.0981683564  0.068152714  0.047345551  0.023749452  0.000000000
#> [46,]  0.0000000000  0.000000000  0.000000000  0.000000000  0.000000000
#> [47,] -0.0746837684 -0.037493931 -0.011183582  0.000000000  0.000000000
#> [48,]  0.0228498761  0.000000000  0.000000000  0.000000000  0.000000000
#> [49,] -0.0552153461 -0.003864964  0.000000000  0.000000000  0.000000000
#> [50,]  0.0754632213  0.074751349  0.059875868  0.033237629  0.002208185
#>               [,6]         [,7]        [,8]         [,9]       [,10]
#>  [1,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>  [2,] -0.086108127 -0.055619146 -0.03049181 -0.013949124  0.00000000
#>  [3,]  0.008211179  0.000000000  0.00000000  0.000000000  0.00000000
#>  [4,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>  [5,] -0.035793788 -0.013207816  0.00000000  0.000000000  0.00000000
#>  [6,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>  [7,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>  [8,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>  [9,] -0.010210172  0.000000000  0.00000000  0.000000000  0.00000000
#> [10,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [11,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [12,] -0.015629647  0.000000000  0.00000000  0.000000000  0.00000000
#> [13,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [14,]  5.100193572  5.100941314  5.10174271  5.101590222  5.10140769
#> [15,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [16,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [17,] -0.113167428 -0.105958902 -0.09752567 -0.085816376 -0.07417618
#> [18,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [19,] -0.094415826 -0.084941812 -0.07491951 -0.061400309 -0.04825775
#> [20,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [21,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [22,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [23,] -0.062113900 -0.044192964 -0.02513273 -0.008364712  0.00000000
#> [24,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [25,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [26,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [27,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [28,]  4.876827891  4.886823121  4.89261173  4.892385032  4.89150591
#> [29,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [30,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [31,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [32,]  0.017039719  0.000000000  0.00000000  0.000000000  0.00000000
#> [33,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [34,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [35,]  0.070119636  0.057101815  0.04412478  0.032432340  0.01995194
#> [36,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [37,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [38,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [39,]  4.763658946  4.775984110  4.78941822  4.793695136  4.79842384
#> [40,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [41,] -0.042341576 -0.017295985  0.00000000  0.000000000  0.00000000
#> [42,] -0.017664555 -0.004312016  0.00000000  0.000000000  0.00000000
#> [43,] -0.003629897  0.000000000  0.00000000  0.000000000  0.00000000
#> [44,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [45,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [46,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [47,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [48,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [49,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#> [50,]  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000
#>              [,11]       [,12]       [,13]       [,14]        [,15]    [,16]
#>  [1,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [2,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [3,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [4,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [5,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [6,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [7,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [8,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>  [9,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [10,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [11,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [12,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [13,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [14,]  5.100477159  5.10010076  5.09437951  5.09342987  5.091599121 5.090068
#> [15,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [16,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [17,] -0.060685586 -0.04747616 -0.04041737 -0.02427318 -0.008095172 0.000000
#> [18,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [19,] -0.035195058 -0.02197292 -0.01638827  0.00000000  0.000000000 0.000000
#> [20,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [21,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [22,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [23,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [24,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [25,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [26,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [27,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [28,]  4.889677318  4.88937584  4.88561241  4.88528314  4.881884615 4.880100
#> [29,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [30,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [31,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [32,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [33,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [34,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [35,]  0.008240951  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [36,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [37,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [38,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [39,]  4.795770089  4.79621402  4.79678217  4.79359557  4.789025559 4.787492
#> [40,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [41,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [42,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [43,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [44,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [45,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [46,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [47,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [48,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [49,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#> [50,]  0.000000000  0.00000000  0.00000000  0.00000000  0.000000000 0.000000
#>          [,17]    [,18]    [,19]    [,20]    [,21]    [,22]    [,23]    [,24]
#>  [1,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [2,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [4,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [5,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [6,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [7,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [8,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [10,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [11,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [12,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [13,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [14,] 5.087687 5.086617 5.084719 5.082613 5.081247 5.079110 5.076969 5.075442
#> [15,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [16,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [17,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [18,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [19,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [20,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [21,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [22,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [23,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [24,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [25,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [26,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [27,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [28,] 4.877219 4.876688 4.874735 4.872398 4.871415 4.869097 4.866781 4.865565
#> [29,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [30,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [31,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [32,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [33,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [34,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [35,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [36,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [37,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [38,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [39,] 4.783269 4.783975 4.781712 4.778538 4.778353 4.775389 4.772429 4.771761
#> [40,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [41,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [42,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [43,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [44,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [45,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [46,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [47,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [48,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [49,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [50,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>          [,25]
#>  [1,] 0.000000
#>  [2,] 0.000000
#>  [3,] 0.000000
#>  [4,] 0.000000
#>  [5,] 0.000000
#>  [6,] 0.000000
#>  [7,] 0.000000
#>  [8,] 0.000000
#>  [9,] 0.000000
#> [10,] 0.000000
#> [11,] 0.000000
#> [12,] 0.000000
#> [13,] 0.000000
#> [14,] 5.073846
#> [15,] 0.000000
#> [16,] 0.000000
#> [17,] 0.000000
#> [18,] 0.000000
#> [19,] 0.000000
#> [20,] 0.000000
#> [21,] 0.000000
#> [22,] 0.000000
#> [23,] 0.000000
#> [24,] 0.000000
#> [25,] 0.000000
#> [26,] 0.000000
#> [27,] 0.000000
#> [28,] 4.864268
#> [29,] 0.000000
#> [30,] 0.000000
#> [31,] 0.000000
#> [32,] 0.000000
#> [33,] 0.000000
#> [34,] 0.000000
#> [35,] 0.000000
#> [36,] 0.000000
#> [37,] 0.000000
#> [38,] 0.000000
#> [39,] 4.770967
#> [40,] 0.000000
#> [41,] 0.000000
#> [42,] 0.000000
#> [43,] 0.000000
#> [44,] 0.000000
#> [45,] 0.000000
#> [46,] 0.000000
#> [47,] 0.000000
#> [48,] 0.000000
#> [49,] 0.000000
#> [50,] 0.000000
#> 
#> $hata
#>               [,1]          [,2]         [,3]         [,4]          [,5]
#>  [1,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#>  [2,]  0.069560562  0.0783505673  0.086684199  0.090633679  0.0881532413
#>  [3,]  0.168871997  0.1587648907  0.144557381  0.130045354  0.1081551331
#>  [4,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#>  [5,]  0.020418375  0.0111570940  0.005988069  0.000276387  0.0000000000
#>  [6,]  0.157633825  0.1277533439  0.090735530  0.048359729  0.0000000000
#>  [7,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#>  [8,] -0.126694378 -0.1194737865 -0.109341284 -0.097986299 -0.0827380153
#>  [9,]  0.019346861  0.0027908869  0.000000000  0.000000000  0.0000000000
#> [10,]  0.104868961  0.0967618595  0.084186352  0.065731899  0.0420593686
#> [11,] -0.100956366 -0.0830150467 -0.070737267 -0.054242311 -0.0421637992
#> [12,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [13,] -0.064664704 -0.0441438325 -0.020413200  0.000000000  0.0000000000
#> [14,]  4.912506847  4.8861550783  4.855505050  4.823788744  4.7882331487
#> [15,] -0.011856292 -0.0057763343 -0.001703663  0.000000000  0.0000000000
#> [16,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [17,]  0.004479541  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [18,] -0.237508538 -0.2359275616 -0.228840181 -0.213656388 -0.1943182496
#> [19,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [20,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [21,]  0.017177252  0.0001402115  0.000000000  0.000000000  0.0000000000
#> [22,]  0.009905589  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [23,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [24,]  0.133864922  0.1198649320  0.109107824  0.099541804  0.0867999594
#> [25,]  0.000000000 -0.0038158486 -0.009524607 -0.008667827 -0.0007774301
#> [26,] -0.171954119 -0.1655185613 -0.162626487 -0.151789951 -0.1309907834
#> [27,]  0.092670787  0.0731760178  0.050004818  0.021416672  0.0000000000
#> [28,]  4.946861658  4.9301441802  4.902062250  4.871757121  4.8366440774
#> [29,]  0.006036276  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [30,] -0.023310335 -0.0252291058 -0.023112719 -0.016014043 -0.0050759598
#> [31,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [32,] -0.020555929  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [33,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [34,] -0.112177530 -0.1128450416 -0.103395255 -0.091870959 -0.0807750632
#> [35,]  0.022813691  0.0015131090  0.000000000  0.000000000  0.0000000000
#> [36,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [37,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [38,] -0.048777117 -0.0331419500 -0.019024701  0.000000000  0.0000000000
#> [39,]  4.855302978  4.8429062157  4.832993242  4.818383294  4.8011211834
#> [40,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [41,]  0.022963787  0.0069280162  0.000000000  0.000000000  0.0000000000
#> [42,] -0.006556446  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [43,]  0.028215170  0.0132139319  0.000000000  0.000000000  0.0000000000
#> [44,]  0.000000000  0.0000000000  0.000000000  0.000000000  0.0000000000
#> [45,]  0.156657549  0.1180868310  0.083874871  0.045695302  0.0065322859
#> [46,] -0.147765827 -0.1280794345 -0.115421212 -0.098560025 -0.0862433505
#> [47,] -0.012279517 -0.0060854807  0.000000000  0.000000000  0.0000000000
#> [48,] -0.113531369 -0.1035660053 -0.093953036 -0.079015946 -0.0564296624
#> [49,] -0.089372124 -0.0783997115 -0.062907459 -0.048084263 -0.0229125676
#> [50,]  0.082746804  0.0727004291  0.052330598  0.034211950  0.0073257767
#>              [,6]         [,7]        [,8]        [,9]       [,10]    [,11]
#>  [1,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#>  [2,]  0.07445643  0.054953411  0.02817218  0.00000000  0.00000000 0.000000
#>  [3,]  0.08319300  0.051585238  0.01350295  0.00000000  0.00000000 0.000000
#>  [4,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#>  [5,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#>  [6,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#>  [7,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#>  [8,] -0.06781281 -0.047418633 -0.01811448  0.00000000  0.00000000 0.000000
#>  [9,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [10,]  0.01938889  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [11,] -0.02588769 -0.004526774  0.00000000  0.00000000  0.00000000 0.000000
#> [12,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [13,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [14,]  4.75805654  4.725761690  4.68404825  4.63249662  4.57353542 4.495524
#> [15,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [16,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [17,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [18,] -0.17706495 -0.155989813 -0.12131540 -0.07994253 -0.03297150 0.000000
#> [19,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [20,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [21,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [22,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [23,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [24,]  0.08733173  0.089696917  0.08486715  0.05778551  0.02075133 0.000000
#> [25,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [26,] -0.09726618 -0.050438874  0.00000000  0.00000000  0.00000000 0.000000
#> [27,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [28,]  4.81611278  4.790133901  4.75059043  4.71347670  4.66591832 4.599937
#> [29,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [30,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [31,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [32,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [33,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [34,] -0.06311347 -0.033497308  0.00000000  0.00000000  0.00000000 0.000000
#> [35,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [36,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [37,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [38,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [39,]  4.77774611  4.742135091  4.69846059  4.65616608  4.59499688 4.518758
#> [40,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [41,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [42,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [43,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [44,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [45,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [46,] -0.07097587 -0.050133189 -0.02209077  0.00000000  0.00000000 0.000000
#> [47,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [48,] -0.02571828  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [49,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#> [50,]  0.00000000  0.000000000  0.00000000  0.00000000  0.00000000 0.000000
#>          [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22]
#>  [1,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [2,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [3,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [4,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [5,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [6,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [7,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [8,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>  [9,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [10,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [11,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [12,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [13,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [14,] 4.391117     0     0     0     0     0     0     0     0     0     0
#> [15,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [16,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [17,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [18,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [19,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [20,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [21,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [22,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [23,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [24,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [25,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [26,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [27,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [28,] 4.507510     0     0     0     0     0     0     0     0     0     0
#> [29,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [30,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [31,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [32,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [33,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [34,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [35,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [36,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [37,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [38,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [39,] 4.420292     0     0     0     0     0     0     0     0     0     0
#> [40,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [41,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [42,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [43,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [44,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [45,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [46,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [47,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [48,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [49,] 0.000000     0     0     0     0     0     0     0     0     0     0
#> [50,] 0.000000     0     0     0     0     0     0     0     0     0     0
#>       [,23] [,24] [,25]
#>  [1,]     0     0     0
#>  [2,]     0     0     0
#>  [3,]     0     0     0
#>  [4,]     0     0     0
#>  [5,]     0     0     0
#>  [6,]     0     0     0
#>  [7,]     0     0     0
#>  [8,]     0     0     0
#>  [9,]     0     0     0
#> [10,]     0     0     0
#> [11,]     0     0     0
#> [12,]     0     0     0
#> [13,]     0     0     0
#> [14,]     0     0     0
#> [15,]     0     0     0
#> [16,]     0     0     0
#> [17,]     0     0     0
#> [18,]     0     0     0
#> [19,]     0     0     0
#> [20,]     0     0     0
#> [21,]     0     0     0
#> [22,]     0     0     0
#> [23,]     0     0     0
#> [24,]     0     0     0
#> [25,]     0     0     0
#> [26,]     0     0     0
#> [27,]     0     0     0
#> [28,]     0     0     0
#> [29,]     0     0     0
#> [30,]     0     0     0
#> [31,]     0     0     0
#> [32,]     0     0     0
#> [33,]     0     0     0
#> [34,]     0     0     0
#> [35,]     0     0     0
#> [36,]     0     0     0
#> [37,]     0     0     0
#> [38,]     0     0     0
#> [39,]     0     0     0
#> [40,]     0     0     0
#> [41,]     0     0     0
#> [42,]     0     0     0
#> [43,]     0     0     0
#> [44,]     0     0     0
#> [45,]     0     0     0
#> [46,]     0     0     0
#> [47,]     0     0     0
#> [48,]     0     0     0
#> [49,]     0     0     0
#> [50,]     0     0     0
#> 
#> $medest
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [2,] -0.0158348201 -1.398371e-02 -1.363375e-02 -1.259425e-02 -1.033730e-02
#>  [3,]  0.0209841305  1.612460e-02  1.163505e-02  7.166705e-03  3.198740e-03
#>  [4,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [5,] -0.0034878270 -1.484315e-03 -6.933901e-04 -2.557949e-05  0.000000e+00
#>  [6,] -0.0012195830  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [7,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [8,]  0.0025137564  9.705612e-04  0.000000e+00  0.000000e+00  0.000000e+00
#>  [9,] -0.0004643435 -1.023307e-04  0.000000e+00  0.000000e+00  0.000000e+00
#> [10,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [11,] -0.0060787625 -1.538171e-03  0.000000e+00  0.000000e+00  0.000000e+00
#> [12,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [13,] -0.0052814918 -2.003196e-03 -1.755942e-04  0.000000e+00  0.000000e+00
#> [14,] 25.1625320163  2.497146e+01  2.473462e+01  2.457677e+01  2.442497e+01
#> [15,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [16,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [17,] -0.0006327977  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [18,] -0.0194593433 -1.692608e-03  0.000000e+00  0.000000e+00  0.000000e+00
#> [19,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [20,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [21,] -0.0003521479 -6.583969e-07  0.000000e+00  0.000000e+00  0.000000e+00
#> [22,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [23,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [24,] -0.0146839665 -8.736633e-03 -5.256817e-03 -2.740879e-03 -7.583310e-04
#> [25,]  0.0000000000 -2.498298e-05 -4.054411e-05  0.000000e+00  0.000000e+00
#> [26,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [27,]  0.0065135734  3.109623e-03  9.574293e-04  3.138156e-05  0.000000e+00
#> [28,] 23.6519093799  2.378286e+01  2.373709e+01  2.364843e+01  2.352975e+01
#> [29,] -0.0004067780  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [30,]  0.0011211511  7.543631e-04  3.225202e-04  0.000000e+00  0.000000e+00
#> [31,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [32,] -0.0029635913  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [33,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [34,]  0.0029035294  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [35,]  0.0043035159  2.299922e-04  0.000000e+00  0.000000e+00  0.000000e+00
#> [36,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [37,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [38,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [39,] 22.7894965332  2.285794e+01  2.290927e+01  2.290906e+01  2.285044e+01
#> [40,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [41,] -0.0046467279 -1.098227e-03  0.000000e+00  0.000000e+00  0.000000e+00
#> [42,]  0.0006577220  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [43,] -0.0029545151 -8.574438e-04  0.000000e+00  0.000000e+00  0.000000e+00
#> [44,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [45,]  0.0153788141  8.047938e-03  3.971102e-03  1.085238e-03  0.000000e+00
#> [46,]  0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [47,]  0.0009170806  2.281686e-04  0.000000e+00  0.000000e+00  0.000000e+00
#> [48,] -0.0025941777  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [49,]  0.0049347128  3.030120e-04  0.000000e+00  0.000000e+00  0.000000e+00
#> [50,]  0.0062443404  5.434455e-03  3.133340e-03  1.137124e-03  1.617667e-05
#>                [,6]         [,7]          [,8]     [,9]    [,10]    [,11]
#>  [1,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [2,] -0.0064113036 -0.003056462 -0.0008590207  0.00000  0.00000  0.00000
#>  [3,]  0.0006831126  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [4,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [5,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [6,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [7,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [8,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>  [9,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [10,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [11,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [12,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [13,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [14,] 24.2670093806 24.105833045 23.8968090454 23.63310 23.33147 22.92932
#> [15,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [16,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [17,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [18,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [19,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [20,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [21,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [22,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [23,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [24,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [25,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [26,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [27,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [28,] 23.4873531229 23.408537099 23.2427944605 23.06014 22.82337 22.49221
#> [29,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [30,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [31,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [32,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [33,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [34,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [35,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [36,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [37,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [38,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [39,] 22.7595529813 22.648361840 22.5028927394 22.32024 22.04874 21.67092
#> [40,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [41,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [42,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [43,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [44,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [45,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [46,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [47,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [48,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [49,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#> [50,]  0.0000000000  0.000000000  0.0000000000  0.00000  0.00000  0.00000
#>          [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22]
#>  [1,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [2,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [3,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [4,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [5,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [6,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [7,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [8,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>  [9,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [10,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [11,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [12,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [13,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [14,] 22.39514     0     0     0     0     0     0     0     0     0     0
#> [15,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [16,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [17,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [18,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [19,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [20,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [21,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [22,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [23,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [24,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [25,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [26,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [27,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [28,] 22.03891     0     0     0     0     0     0     0     0     0     0
#> [29,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [30,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [31,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [32,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [33,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [34,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [35,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [36,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [37,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [38,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [39,] 21.20067     0     0     0     0     0     0     0     0     0     0
#> [40,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [41,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [42,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [43,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [44,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [45,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [46,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [47,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [48,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [49,]  0.00000     0     0     0     0     0     0     0     0     0     0
#> [50,]  0.00000     0     0     0     0     0     0     0     0     0     0
#>       [,23] [,24] [,25]
#>  [1,]     0     0     0
#>  [2,]     0     0     0
#>  [3,]     0     0     0
#>  [4,]     0     0     0
#>  [5,]     0     0     0
#>  [6,]     0     0     0
#>  [7,]     0     0     0
#>  [8,]     0     0     0
#>  [9,]     0     0     0
#> [10,]     0     0     0
#> [11,]     0     0     0
#> [12,]     0     0     0
#> [13,]     0     0     0
#> [14,]     0     0     0
#> [15,]     0     0     0
#> [16,]     0     0     0
#> [17,]     0     0     0
#> [18,]     0     0     0
#> [19,]     0     0     0
#> [20,]     0     0     0
#> [21,]     0     0     0
#> [22,]     0     0     0
#> [23,]     0     0     0
#> [24,]     0     0     0
#> [25,]     0     0     0
#> [26,]     0     0     0
#> [27,]     0     0     0
#> [28,]     0     0     0
#> [29,]     0     0     0
#> [30,]     0     0     0
#> [31,]     0     0     0
#> [32,]     0     0     0
#> [33,]     0     0     0
#> [34,]     0     0     0
#> [35,]     0     0     0
#> [36,]     0     0     0
#> [37,]     0     0     0
#> [38,]     0     0     0
#> [39,]     0     0     0
#> [40,]     0     0     0
#> [41,]     0     0     0
#> [42,]     0     0     0
#> [43,]     0     0     0
#> [44,]     0     0     0
#> [45,]     0     0     0
#> [46,]     0     0     0
#> [47,]     0     0     0
#> [48,]     0     0     0
#> [49,]     0     0     0
#> [50,]     0     0     0
#> 
#> $alpha
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $tau
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $lambda
#>  [1] 0.01980263 0.03922071 0.05826891 0.07696104 0.09531018 0.11332869
#>  [7] 0.13102826 0.14842001 0.16551444 0.18232156 0.19885086 0.21511138
#> [13] 0.23111172 0.24686008 0.26236426 0.27763174 0.29266961 0.30748470
#> [19] 0.32208350 0.33647224 0.35065687 0.36464311 0.37843644 0.39204209
#> [25] 0.40546511
#> 
#> $nump
#>  [1] 78 70 58 47 39 31 24 18 14 12 10  9  6  5  5  4  4  4  4  4  4  4  4  4  4
#>